We fixed the bot's buttons, fixed the API for user registration, and fixed the database model on big int

+++ backend/src/infrastructure/db/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, Text, Enum as SAEnum
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, Text, Enum as SAEnum, BigInteger
 from sqlalchemy.orm import relationship, declarative_base
 from datetime import datetime
 from domain.domain_models import BookStatus
@@ -9,7 +9,7 @@ Base = declarative_base()
 class UserTable(Base):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True, index=True)  # Telegram ID
+    id = Column(BigInteger, primary_key=True, index=True)  # Telegram ID
     full_name = Column(String, nullable=False)
     username = Column(String, nullable=True, index=True)  # Индекс для поиска
     is_admin = Column(Boolean, default=False, index=True)  # Индекс для фильтрации админов
@@ -26,14 +26,14 @@ class BookTable(Base):
     genre = Column(String(50), nullable=True, index=True)  # Индекс для фильтрации
     isbn = Column(String(20), nullable=True)
 
-    owner_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
+    owner_id = Column(BigInteger, ForeignKey("users.id"), nullable=False, index=True)
     owner = relationship("UserTable", foreign_keys=[owner_id], backref="owned_books")
 
     status = Column(SAEnum(BookStatus), default=BookStatus.AVAILABLE, index=True)  # Индекс для фильтрации
     image_path = Column(String, nullable=True)
 
     # Логика выдачи
-    borrower_id = Column(Integer, ForeignKey("users.id"), nullable=True, index=True)
+    borrower_id = Column(BigInteger, ForeignKey("users.id"), nullable=True, index=True)
     borrower = relationship("UserTable", foreign_keys=[borrower_id], backref="borrowed_books")
 
     return_due_date = Column(DateTime, nullable=True, index=True)  # Индекс для проверки просрочек
@@ -47,7 +47,7 @@ class BookHistoryTable(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     book_id = Column(Integer, ForeignKey("books.id"), nullable=False, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False)  # Кто совершил действие
+    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False)  # Кто совершил действие
     status_to = Column(SAEnum(BookStatus), nullable=False)
     comment = Column(String, nullable=True)
     photo_proof_path = Column(String, nullable=True)
@@ -64,7 +64,7 @@ class WaitlistTable(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     book_id = Column(Integer, ForeignKey("books.id"), nullable=False, index=True)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
+    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False, index=True)
     created_at = Column(DateTime, default=datetime.utcnow)
 
     # Уникальный индекс, чтобы пользователь не мог дважды попасть в waitlist одной книги
